Add required libstdc++6 setup to Travis config
The Travis build environment fails to execute the "extended" version
of Hugo, leading to "libstdc++.so.6:version `GLIBCXX_3.4.20' not
found" errors (for 3.4.20 and 3.4.21).

Evidently this is a known issue on Hugo's forums, as seen here:
https://discourse.gohugo.io/t/solved-hugo-v44-extended-and-relocation-errors-on-travis/13029

Per that thread, we have to manually install the libstdc++6 from
Xenial (rather than Trusty's version) for the Hugo extended version
to run properly, allowing the integration test for extended to pass.